Create `ParseWorldStatusPage` function
This function is exported only because we want to test it. When
possible, it should be unexported.

With the current implementation it's not possible to move it to an
`internal` package because it would result in an import cycle.

The cleanest option would probably be using a custom `*http.Client` that
returns the HTML from `testdata/worldstatus.html` when the expected URL
is requested (and an error otherwise).